Add debugging commands and fix mate blunder bug
Bug cause: when I was integrating transposition tables I moved the
transition to quiescence search above the mate and draw checks, since
those require the legal moves to be generated, and I was thinking "the
quiescence search already needs to generate the moves, I don't need to
do it again". However, the quiescence search doesn't do mate-finding,
and so mate-in-1s can become invisible to the engine and cause it to
just blunder mate sometimes!